# swagger/api.yaml
openapi: "3.0.0"
info:
  title: BG Library
  version: "1.0"
servers:
  - url: ${API_URL}
    description: Dynamic Server
paths:
...
  /api/v1/library/game/id/{gameId}:
    get:
      tags:
        - games
      operationId: getGame
      summary: Get a game
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      responses:
        '200':
          description: Game found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - games
      operationId: updateGame
      summary: Update an existing game
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '204':
          description: Game successfully updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - games
      operationId: deleteGame
      summary: Soft delete a game
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      responses:
        '204':
          description: Game deleted successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/library/checkin:
    post:
      tags:
        - transactions
      operationId: checkInGame
      summary: Check in a game
      description: |
        Make an idempotent call to check in a game.
        If the game is already checked in, the most recent library transaction will be returned will be returned
      parameters:
        - in: query
          name: transactionId
          schema:
            type: string
            format: uuid
          required: true
          description: Transaction ID
      responses:
        '204':
          description: Check in successful
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/checkout:
    post:
      tags:
        - transactions
      operationId: checkOutGame
      summary: Check out a game
      description: |
        Make an idempotent call to check out a game.
        If the game is already checked out, the existing checkout will be returned
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckOutRequest'
      responses:
        '201':
          description: Check out successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryTransaction'
        '400':
          description: Game/Patron not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Game already checked out to a different patron
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/transactions:
    get:
      tags:
        - transactions
      operationId: listTransactionEvents
      summary: List transaction events
      description: |
        List all library transactions, ordered by timestamp (newest first).
        Game titles and patron names filters stack, so if you filter by both title and patron name, 
        only games with that title and patron name will be returned
      parameters:
        - in: query
          name: game_title
          schema:
            type: string
          description: Filter by game title (optional)
        - in: query
          name: patron_name
          schema:
            type: string
          description: Filter by patron name (optional)
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          description: Limit the number of transactions returned (optional)
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Offset for pagination (optional)
      responses:
        '200':
          description: List of library transactions successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionEventList'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ptw/game/gameId/{gameId}:
    post:
      tags:
        - ptw
        - games
      operationId: addPlayToWinGame
      summary: Mark a library game as play to win
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      responses:
        '204':
          description: Game marked as play to win successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - ptw
        - games
      operationId: removePlayToWinGame
      summary: Mark a library game as not play to win
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemovePlayToWinGameRequest'
      responses:
        '204':
          description: Game marked as not play to win successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ptw/session:
    post:
      tags:
        - ptw
      operationId: addPlayToWinSession
      summary: Add a play to win session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayToWinSessionRequest'
      responses:
        '201':
          description: Play to win session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayToWinSession'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Play to win game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ptw/entries/playToWinId/{playToWinId}:
    get:
      tags:
        - ptw
      operationId: getPlayToWinSessionEntries
      parameters:
        - in: path
          name: playToWinId
          schema:
            type: string
            format: uuid
          required: true
          description: Play to win game ID
      description: Get play to win entries by game
      summary: Get all play to win entries for a play to win game
      responses:
        '200':
          description: Play to win entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayToWinEntryList'
components:
  schemas:
...
    Game:
      type: object
      required:
        - gameId
        - title
        - isPlayToWin
      properties:
        gameId:
          type: string
          format: uuid
          description: Game ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the board game
          example: "Catan"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the game
          example: "9780307455925"
        isPlayToWin:
          type: boolean
          description: Indicates if the game is a play-to-win game
          example: true
          default: false
      description: A game in the library
...
    CreateGameRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the board game
          example: "Catan"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the game
          example: "9780307455925"
        isPlayToWin:
          type: boolean
          description: Indicates if the game is a play-to-win game
          example: true
          default: false
      description: Request payload for creating a new game
    Patron:
      type: object
      required:
        - patronId
        - name
      properties:
        patronId:
          type: string
          format: uuid
          description: Patron ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the patron
          example: "John Smith"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the patron
          example: "9780307455925"
      description: A library patron
...
    CheckOutRequest:
      type: object
      required:
        - patronId
        - gameId
      properties:
        patronId:
          type: string
          format: uuid
          description: Patron ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        gameId:
          type: string
          format: uuid
          description: Game ID
          example: "cc4bda14-ee18-45c0-b09b-a1dcd18c0df9"
      description: Request payload for making a check in/out transaction
    LibraryTransaction:
      type: object
      required:
        - id
        - patronId
        - gameId
        - timestamp
      properties:
        id:
          type: string
          format: uuid
          description: Transaction ID
          example: "f13bb82c-18dd-4a94-a878-44ca6706e2d9"
        patronId:
          type: string
          format: uuid
          description: Patron ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        gameId:
          type: string
          format: uuid
          description: Game ID
          example: "cc4bda14-ee18-45c0-b09b-a1dcd18c0df9"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the transaction occurred in UTC
          example: "2026-01-31T12:00:00Z"
      description: Request payload for making a check in/out transaction
    TransactionEvent:
      type: object
      required:
        - transaction_id
        - patron
        - game
        - event_timestamp
        - event_type
      properties:
        transaction_id:
          type: string
          format: uuid
          description: Transaction ID
          example: "f13bb82c-18dd-4a94-a878-44ca6706e2d9"
        game:
          $ref: "#/components/schemas/Game"
        patron:
          $ref: "#/components/schemas/Patron"
        event_timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the transaction occurred in UTC
          example: "2026-01-31T12:00:00Z"
        event_type:
          type: string
          enum:
            - check_in
            - check_out
          description: Type of transaction event
          example: "check_out"
      description: Record of a library transaction event
    TransactionEventList:
      type: object
      required:
        - transactions
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionEvent'
          description: Array of transaction events
      description: List of transaction events
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
                - NOT_FOUND
                - CONFLICT
                - MALFORMED_REQUEST
              description: Machine readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Validation Error"
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
              description: Array of specific field errors. Empty if no specific errors
      description: Response when errors occur
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: The input field that caused the error
          example: "title"
        message:
          type: string
          description: Human-readable error message for this field
          example: "Title cannot exceed 100 characters"
      description: Details about specific errors
...
    RemovePlayToWinGameRequest:
      type: object
      required:
        - RemovalReason
      properties:
        RemovalReason:
          type: string
          enum:
            - claimed
            - mistake
            - other
          description: The reason for removing the game from play to win
          example: "mistake"
        RemovalComment:
          type: string
          description: Optional comment explaining the reason for removing the game from play to win
          example: "The game was a library copy, not a play to win game"
          maxLength: 500
      description: Request payload for removing a game from play to win
    CreatePlayToWinSessionRequest:
      type: object
      required:
        - playToWinId
        - entries
      properties:
        playToWinId:
          type: string
          format: uuid
          description: The play to win ID this session is tied to
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        playtimeMinutes:
          type: integer
          description: The estimated playtime for this session in minutes
          example: 60
          minimum: 0
        entries:
          type: array
          items:
            type: object
            required:
              - entrantName
              - entrantUniqueId
            properties:
              entrantName:
                type: string
                maxLength: 100
                minLength: 1
                example: "John Smith"
                description: The name of the entrant
              entrantUniqueId:
                type: string
                maxLength: 100
                minLength: 1
                example: "1234567890"
                description: The unique ID of the entrant
            description: An entry in the play to win session
          description: Array of entries in the play to win session
      description: Request payload for creating a play to win session
    PlayToWinEntry:
      type: object
      required:
        - entryId
        - entrantName
        - entrantUniqueId
      properties:
        entryId:
          type: string
          format: uuid
          description: The ID of the entry
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        entrantName:
          type: string
          maxLength: 100
          minLength: 1
          example: "John Smith"
          description: The name of the entrant
        entrantUniqueId:
          type: string
          maxLength: 100
          minLength: 1
          example: "1234567890"
          description: The unique ID of the entrant
      description: A single play to win entry
    PlayToWinSession:
      type: object
      required:
        - sessionId
        - playToWinEntries
      properties:
        sessionId:
          type: string
          format: uuid
          description: The ID of the session
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        playtimeMinutes:
          type: integer
          format: int32
          minLength: 0
          example: 60
          description: The estimated playtime for this session in minutes
        playToWinEntries:
          type: array
          items:
            $ref: '#/components/schemas/PlayToWinEntry'
          description: Array of entries in the play to win session
      description: A play to win game session
    PlayToWinEntryList:
      type: object
      required:
        - entries
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PlayToWinEntry'
          description: Array of play to win entries
      description: List of play to win entries